complete refactoring of powerups
