+++ src/components/Game/Question/Powerups.js
@@ -24,8 +24,8 @@ const powerUpsArr = [
     event: "half",
   },
   {
-    text: "⌛⁺",
-    event: "extraTime",
+    text: "⟳",
+    event: "resetTimer",
   },
   {
     text: "❄️",
