+++ src/components/Game/Question/index.js
@@ -4,17 +4,13 @@ import { useQuiz } from "../../../context/quiz-context";
 import questions from "../../../constants/questions";
 import Stats from "./Stats";
 import Options from "./Options";
-import { useTimer } from "./question.hooks";
-import { useState, useEffect, useMemo } from "react";
+import Timer from "./Timer";
+
 const QuestionNO = styled.div`
   display: inline-block;
   padding: 0.2em 0.5em;
 `;
-const Timing = styled.div`
-  text-align: end;
-  float: right;
-  padding: 0.2em 0.5em;
-`;
+
 const QuestionWrapper = styled.div`
   padding: 1em;
   min-height: 15vh;
@@ -26,20 +22,10 @@ const QuestionH = styled.h2`
 
 const Question = () => {
   const {
-    quizState: {
-      questionIndex,
-      enabledPowerUps: { freeze },
-    },
+    quizState: { questionIndex },
     theme,
   } = useQuiz();
   const currentQuestion = questions[questionIndex];
-  // const timeLeft = useTimer(currentQuestion.time);
-
-  const {
-    quizState: { displayResult },
-    quizDispatch,
-  } = useQuiz();
-  const timeLeft = useTimer(currentQuestion.time);
 
   const CardWrapper = styled.div`
     border: 3px solid ${theme};
@@ -56,7 +42,7 @@ const Question = () => {
           <QuestionNO>
             {questionIndex + 1} of {questions.length}
           </QuestionNO>
-          <Timing> {secondsToMMSS(timeLeft)}</Timing>
+          <Timer />
         </CardHeaderWrapper>
         <QuestionWrapper>
           <QuestionH>
@@ -71,9 +57,3 @@ const Question = () => {
 };
 
 export default Question;
-
-const secondsToMMSS = (seconds) => {
-  var m = Math.floor(seconds / 60);
-  var s = seconds % 60;
-  return (m < 10 ? "0" : "") + m + ":" + ("0" + s).slice(-2);
-};
