+++ src/components/Start/Rules.js
@@ -104,7 +104,7 @@ const Rules = ({ startGame }) => {
           ½ <Span>reduce options to 2</Span>
         </GridChild>
         <GridChild>
-          ⌛⁺ <Span>get 50% extra of the time</Span>
+          ⟳ <Span>Reset Timer</Span>
         </GridChild>
         <GridChild>
           ❄️ <Span>freeze the timer</Span>
